third_party/WebKit/LayoutTests/inspector-protocol/css/css-get-platform-fonts.html

<html>
<head>
<script type="text/javascript" src="../../http/tests/inspector-protocol/inspector-protocol-test.js"></script>
<script type="text/javascript" src="../../http/tests/inspector-protocol/css-protocol-test.js"></script>
<script type="text/javascript" src="../../http/tests/inspector-protocol/dom-protocol-test.js"></script>
<script type="text/javascript">

function test()
{
    var documentNodeId;

    InspectorTest.requestDocumentNodeId(onDocumentNodeId);

    function onDocumentNodeId(nodeId)
    {
        documentNodeId = nodeId;
        InspectorTest.runTestSuite([
            function testFirstLetterPseudoClass(next)
            {
                platformFontsForElementWithSelector("#fancy", next);
            },

            function testSelectElementPlatformFonts(next)
            {
                platformFontsForElementWithSelector("select", next);
            }
        ]);
    }


    function platformFontsForElementWithSelector(selector, callback)
    {
        InspectorTest.requestNodeId(documentNodeId, selector, onNodeId);

        function onNodeId(nodeId)
        {
            InspectorTest.sendCommandOrDie("CSS.getPlatformFontsForNode", { nodeId: nodeId }, onGotComputedFonts);
        }

        function onGotComputedFonts(response)
        {
            dumpComputedFonts(response);
            callback();
        }
    }

    function dumpComputedFonts(response)
    {
        var fonts = response.fonts;
        fonts.sort(function(a, b) {
            return b.glyphCount - a.glyphCount;
        });
        for (var i = 0; i < fonts.length; ++i)
            fonts[i].familyName = "<Some-family-name-" + i + ">";
        InspectorTest.log(JSON.stringify(fonts));
    }
};

window.addEventListener("DOMContentLoaded", function () {
    runTest();
}, false);

</script>
<style>
    #fancy {
        font-family: 'Arial';
        background-color: gray;
    }
    #fancy:first-letter {
        font-family: 'Times New Roman';
        font-size: 400%;
        background-color: blue;
    }

    #fancy:first-line {
        font-family: 'Courier New';
        background-color: yellow;
    }
</style>
</head>
<body>

<div id="fancy">
7chars.<br>
Some line with 29 characters.
</div>
<select>
    <option>Short</option>
    <option selected>Option with a lot of chars.</option>
</select>

</body>
</html>
